+++ paper.bib
@@ -63,6 +63,15 @@ journal = {​GitHub repository},
 howpublished = {​https://github.com/spatialaudio/python-sofa}
 }
 
+@misc{unitful,
+author = {Andrew Keller},
+title = {Unitful.jl},
+year = {2016},
+publisher = {​GitHub},
+journal = {​GitHub repository},
+howpublished = {​https://github.com/PainterQubits/Unitful.jl}
+}
+
 @misc{sampled-signals,
 author = { Spencer Russell },
 title = {SampledSignals.jl},
