### 2021-08-07 744f45d
Simplify unit handling for modulation modifier
