### 2021-08-08 72e9acf
Further simplification
