+++ test/runtests.jl
@@ -516,22 +516,22 @@ end
                     @test mid_mod > end_mod
                 
                     # Test different ways of instanciating the modifier
-                    @test AmplitudeModulation(10u"Hz").rate == 10
-                    @test AmplitudeModulation(1.0u"Hz").rate == 1
-                    @test AmplitudeModulation(10u"Hz", 0.0).rate == 10
-                    @test AmplitudeModulation(10.3u"Hz", 0.0).rate == 10.3
-                    @test AmplitudeModulation(10u"Hz", π).rate == 10
-                    @test AmplitudeModulation(10u"Hz", π, 0.0).rate == 10
-                    @test AmplitudeModulation(10u"Hz", π, 0.5).rate == 10
-                    @test AmplitudeModulation(10.8u"Hz", π, 0.5).rate == 10.8
-                    @test AmplitudeModulation(10u"Hz", π, 1.5).rate == 10
+                    @test AmplitudeModulation(10u"Hz").rate == 10u"Hz"
+                    @test AmplitudeModulation(1.0u"Hz").rate == 1u"Hz"
+                    @test AmplitudeModulation(10u"Hz", 0.0).rate == 10u"Hz"
+                    @test AmplitudeModulation(10.3u"Hz", 0.0).rate == 10.3u"Hz"
+                    @test AmplitudeModulation(10u"Hz", π).rate == 10u"Hz"
+                    @test AmplitudeModulation(10u"Hz", π, 0.0).rate == 10u"Hz"
+                    @test AmplitudeModulation(10u"Hz", π, 0.5).rate == 10u"Hz"
+                    @test AmplitudeModulation(10.8u"Hz", π, 0.5).rate == 10.8u"Hz"
+                    @test AmplitudeModulation(10u"Hz", π, 1.5).rate == 10u"Hz"
                     @test AmplitudeModulation(10u"Hz", π, 1.5).depth == 1.5
-                    @test AmplitudeModulation(rate=3).rate == 3
+                    @test AmplitudeModulation(rate=3u"Hz").rate == 3.0u"Hz"
                     @test AmplitudeModulation(phase=3).phase == 3
                     @test AmplitudeModulation(depth=0.5).depth == 0.5
 
                     # Test defaults
-                    @test AmplitudeModulation().rate == 0
+                    @test AmplitudeModulation().rate == 0u"Hz"
                     @test AmplitudeModulation().phase == π
                     @test AmplitudeModulation().depth == 1
                     @test AmplitudeModulation().enable == true
