### 2021-08-08 c3b3698
Stricter typing
