+++ test/runtests.jl
@@ -516,6 +516,9 @@ end
                     @test mid_mod > end_mod
                 
                     # Test different ways of instanciating the modifier
+                    @test AmplitudeModulation(1u"MHz").rate == 1000000u"Hz"
+                    @test AmplitudeModulation(1u"kHz").rate == 1000u"Hz"
+                    @test AmplitudeModulation(1u"mHz").rate == 0.001u"Hz"
                     @test AmplitudeModulation(10u"Hz").rate == 10u"Hz"
                     @test AmplitudeModulation(1.0u"Hz").rate == 1u"Hz"
                     @test AmplitudeModulation(10u"Hz", 0.0).rate == 10u"Hz"
@@ -524,7 +527,8 @@ end
                     @test AmplitudeModulation(10u"Hz", π, 0.0).rate == 10u"Hz"
                     @test AmplitudeModulation(10u"Hz", π, 0.5).rate == 10u"Hz"
                     @test AmplitudeModulation(10.8u"Hz", π, 0.5).rate == 10.8u"Hz"
-                    @test AmplitudeModulation(10u"Hz", π, 1.5).rate == 10u"Hz"
+                    @test AmplitudeModulation(1.0u"kHz", π, 1.5).rate == 1000u"Hz"
+                    @test AmplitudeModulation(1.0u"kHz", π, 1.5, false).enable == false
                     @test AmplitudeModulation(10u"Hz", π, 1.5).depth == 1.5
                     @test AmplitudeModulation(rate=3u"Hz").rate == 3.0u"Hz"
                     @test AmplitudeModulation(phase=3).phase == 3
