### 2021-08-08 56ba1dd
Simplify unit handling for modulation modifier (#47)
