+++ src/AuditoryStimuli.jl
@@ -12,7 +12,9 @@ using Logging
 using FFTW
 using Unitful
 using Plots
+using Parameters
 
+using Unitful: AbstractQuantity, AffineQuantity, DimensionlessQuantity
 import SampledSignals: nchannels, samplerate, unsafe_read!
 import Plots: plot
 
