### 2021-08-08 564739f
Update Amplification documentation.
