### 2021-08-08 0e4bcea
Update Amplification.jl
