### 2021-08-08 18d3e4c
Update Amplification.jl
