### 2021-08-08 7ee2d30
Update Amplification.jl
