### 2021-08-08 07fd288
Update Amplification documentation. (#49)
