+++ .github/workflows/docs.yml
@@ -9,6 +9,7 @@ on:
       - 'README.md'
     tags: '*'
   pull_request:
+  workflow_dispatch:
 
 jobs:
   build:
