+++ .github/workflows/runtests.yml
@@ -4,7 +4,7 @@ on:
   pull_request:
   push:
     branches:
-      - master
+      - main
     tags: '*'
 
 
