+++ docs/src/example-hs.md
@@ -20,7 +20,7 @@ source = SinusoidSource(Float64, 48u"kHz", stack_frequencies)
 amp = Amplification(current=1/length(stack_frequencies),
                     target=1/length(stack_frequencies),
                     change_limit=1)
-am = AmplitudeModulation(15)
+am = AmplitudeModulation(15u"hz")
 sink = DummySampleSink(Float64, 48u"kHz", 1)
 
 # Run real time audio processing
@@ -47,7 +47,7 @@ using DisplayAs # hide
 
 stack_frequencies = 200:200:2400
 source = SinusoidSource(Float64, 48000, stack_frequencies)
-am = AmplitudeModulation(15)
+am = AmplitudeModulation(15u"hz")
 
 audio = read(source, 1.0u"s") 
 modulated_audio = modify(am, audio) 
