+++ docs/make.jl
@@ -26,4 +26,5 @@ makedocs(
 deploydocs(
     repo = "github.com/rob-luke/AuditoryStimuli.jl.git",
     push_preview = true,
+    devbranch = "main"
 )
