### 2021-08-08 #3582 @vahid-sb
Hi @rob-luke. My pleasure to go through your code and make any suggestions to improve the code and documentation, if I can. but please note that as I told @bmcfee, for work-related reasons, it might be a bit slow, but hopefully not too slow.
