### 2021-08-09 2c3bdc3
CompatHelper: add new compat entry for "Parameters" at version "0.12" (#50)
