+++ Project.toml
@@ -21,6 +21,7 @@ Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 DSP = "0.6.0, 0.7"
 FFTW = "1.1.0"
 Images = "0.23, 0.24"
+Parameters = "0.12"
 Pipe = "1.3.0"
 Plots = "1.0.0"
 SampledSignals = "2.0.0"
