### 2021-08-09 8bc2bea
use Parameters.jl for amplification
