### 2021-08-09 cec512e
Update Amplification.jl
