### 2021-08-09 906f744
Use Parameters.jl for amplification modifier (#51)
