### 2021-08-13 5a2f5aa
Add function to plot cross correlation
