### 2021-08-13 #3613 @whedon
[REVIEW]: AuditoryStimuli.jl: A Julia package for generating real-time auditory stimuli
**Submitting author:** @rob-luke (<a href="http://orcid.org/0000-0002-4930-8351">Robert Luke</a>)
**Repository:** <a href="https://github.com/rob-luke/AuditoryStimuli.jl" target ="_blank">https://github.com/rob-luke/AuditoryStimuli.jl</a>
**Version:** v0.0.11
