+++ test/runtests.jl
@@ -599,6 +599,7 @@ end
             itd = TimeDelay(delay=33, buffer=zeros(33, 1))
             itd = TimeDelay(channel=33)
             itd = TimeDelay(enable=false, channel=3)
+            itd = TimeDelay(2, 0.5u"ms", samplerate=48u"kHz")
 
             # Test correct behaiour
             for desired_itd = -100:10:100
