+++ .github/workflows/wordlist.txt
@@ -64,6 +64,9 @@ TFilter
 th
 TimeDelay
 Undurraga
+Benedikt
+Grothe
+Pecka
 Unitful
 wav
 wavwrite
