### 2021-08-13 2543bc9
Add function to plot cross correlation (#54)
