+++ test/runtests.jl
@@ -241,6 +241,11 @@ ENV["JULIA_DEBUG"] = "all"
         p = plot(sink)
         @test isa(p, Plots.Plot) == true
 
+        p = plot_cross_correlation(sink)
+        @test isa(p, Plots.Plot) == true
+        p = plot_cross_correlation(sink, lags=0.1u"s")
+        @test isa(p, Plots.Plot) == true
+
         end
     end
 end
@@ -594,6 +599,8 @@ end
             itd = TimeDelay(delay=33, buffer=zeros(33, 1))
             itd = TimeDelay(channel=33)
             itd = TimeDelay(enable=false, channel=3)
+            itd = TimeDelay(2, 0.5u"ms", samplerate=48u"kHz")
+            itd = TimeDelay(2, 0.48u"ms", samplerate=48u"kHz")
 
             # Test correct behaiour
             for desired_itd = -100:10:100
