+++ .github/workflows/wordlist.txt
@@ -15,6 +15,7 @@ crosscor
 designmethod
 Deun
 DF
+samplerate
 digitalfilter
 DisplayAs
 DSP
@@ -63,6 +64,9 @@ TFilter
 th
 TimeDelay
 Undurraga
+Benedikt
+Grothe
+Pecka
 Unitful
 wav
 wavwrite
