### 2021-08-14 1a0840f
Add units to harmonic stack example
