### 2021-08-14 9aafbb0
CompatHelper: add new compat entry for UnitfulRecipes at version 1, (keep existing compat) (#56)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
