+++ Project.toml
@@ -28,6 +28,7 @@ Plots = "1.0.0"
 SampledSignals = "2.0.0"
 StatsBase = "0.32.0, 0.33"
 Unitful = "0.17.0, 1.6, 1.7"
+UnitfulRecipes = "1"
 julia = "1.5"
 
 [extras]
