### 2021-08-14 56db23e
Update example-hs.md
