### 2021-08-15 aa743a8
Run tests every Monday
