+++ .github/workflows/runtests.yml
@@ -7,7 +7,7 @@ on:
       - main
     tags: '*'
   schedule:
-    - cron: "1 13 * * 1"
+    - cron: "1 1 * * 1"
 
 
 jobs:
