### 2021-08-19 #3613 @jongwook
Great work @rob-luke! Below are my reviews and suggestions for minor revisions.

- The paper is written very well, succinctly and effectively conveying how the AuditoryStimuli.jl package can be a valuable addition as a real-time auditory signal synthesis tool, where it bridges the efficiency required for real-time applications and the ease of use of Julia programming language.
- Some of the URLs in the citation stick out of the margin and overflow out of the page. I didn't see this issue in a few of JOSS paper I've checked, but I'm not sure this only happens in the draft mode.
- The documentation is very well-written as well, complete with installation instructions, introductory usage, and detailed examples for the various features that the package provides. The documentation website is automatically built and deployed by Documenter.jl.
- Quite a shameless plug, but being a Julia library on audio, [MusicProcessing.jl](https://github.com/JuliaMusic/MusicProcessing.jl) could be considered as a citation to mention alongside librosa. The [publication](https://www.aes.org/e-lib/browse.cfm?elib=18769) is not open-access, but happy to share the preprint.
- The repository is quite mature, with about 3 years in development, and every commit is validated with automated tests, which currently shows 100% code coverage. The Julia code files amount to about 1700 lines of code, ~40% of which is for tests.
- The git repository takes up 200M after the initial clone, which doesn't seem to be used in the `HEAD` of any of the branches. In case those are taken up by a few large files, stripping them from the Git history using `git filter-branch` commands would be helpful
- The GitHub repository currently lacks the [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines), which is one of the JOSS review criteria. It could be a few sentences in the `README.md` welcoming contribution, and addition of `CONTRIBUTING.md` or issue templates would be a great plus, which can show up when someone opens up an issue.
