### 2021-08-19 eeaf167
Create CONTRIBUTING.md
