+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,30 @@
+---
+name: Bug report
+about: Tell us about broken, incorrect, or inconsistent behavior.
+title: ''
+labels: BUG
+assignees: ''
+---
+
+#### Describe the bug
+*Replace this text with a description of the bug.*
+
+
+#### Steps to reproduce
+*Replace this text with a code snippet or minimal working example [MWE] to
+replicate your problem.
+
+[MWE]: https://en.wikipedia.org/wiki/Minimal_Working_Example
+
+
+#### Expected results
+*Replace this text with a description of what you expected to happen.*
+
+
+#### Actual results
+*Replace this text with the actual output, traceback, screenshot, or other
+description of the results.*
+
+
+#### Additional information
+*Replace this text with the output of `using Pkg; Pkg.status()`.*
