+++ .github/ISSUE_TEMPLATE/documentation.md
@@ -1,0 +1,15 @@
+---
+name: Documentation
+about: Request a new tutorial, an improvement to an existing tutorial, or a new term in our glossary.
+title: ''
+labels: DOC
+assignees: ''
+---
+
+*Please make the issue title a succinct description of your proposed change.
+For example: "add 'foo' to the glossary" or "add my favorite visualization to
+the signal in noise example".*
+
+#### Proposed documentation ehancement
+*Describe your proposed enhancement in detail. If you are requesting a new
+glossary term, please include a proposed definition.*
