+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,27 @@
+---
+name: Feature request
+about: Suggest additions or improvements to MNE-Python.
+title: ''
+labels: ENH
+assignees: ''
+---
+
+#### Describe the new feature or enhancement
+*Please provide a clear and concise description of what you want to add or
+change.*
+
+
+#### Describe your proposed implementation
+*Describe how you think the feature or improvement should be implemented (e.g., 
+as a new type on an existing type? as new capability added to an existing
+method?) If you're not sure, please delete this section and the next section.*
+
+
+#### Describe possible alternatives
+*If you've suggested an implementation above, list here any alternative
+implementations you can think of, and brief comments explaining why the chosen
+implementation is better.*
+
+
+#### Additional comments
+*Add any other context or screenshots about the feature request here.*
