+++ paper.bib
@@ -167,3 +167,10 @@ Url = {http://www.sciencedirect.com/science/article/pii/S1053811918300338},
 Volume = {172},
 Year = {2018}}
 
+@inproceedings{kim2017fast,
+title={Fast music and audio processing using the Julia language},
+author={Kim, Jong Wook and Russell, Spencer and Bello, Juan},
+booktitle={Audio Engineering Society Conference: 2017 AES International Conference on Semantic Audio},
+year={2017},
+organization={Audio Engineering Society}
+}
