### 2021-08-19 #3613 @rob-luke
Thank you for the review and kind words @jongwook. I will briefly respond to your feedback. Please let me know if you want more information or further discussion. I also welcome any additional feedback once the review is done too.

> Some of the URLs in the citation stick out of the margin and overflow out of the page. I didn't see this issue in a few of JOSS paper I've checked, but I'm not sure this only happens in the draft mode.

I confirm that I see this too. I looked up the three most recent JOSS papers and I confirm that they do not have this URL formatting problem. I then looked up their .bib files and they look the same as mine (see figure below, the URL string is very long). So I am not sure how to solve this @bmcfee do you have any suggestions as to how this may be solved?

Figure from Vaidya et al., (2021). JPhyloRef: a tool for testing and resolving phyloreferences. Journal of Open Source Software, 6(64), 3374, https://doi.org/10.21105/joss.03374
![image](https://user-images.githubusercontent.com/748691/130060503-fe7faf51-f4d7-4f4f-88f5-4b8a0f847030.png)

> Quite a shameless plug, but being a Julia library on audio, MusicProcessing.jl could be considered as a citation to mention alongside librosa. The publication is not open-access, but happy to share the preprint.

A completely appropriate plug and I should have added this earlier as I am aware of the great package, but didnt know it had an associated article. I have updated the manuscript to include this reference.

> 40% of which is for tests

Yep, I love tests.

> The git repository takes up 200M after the initial clone, which doesn't seem to be used in the HEAD of any of the branches. In case those are taken up by a few large files, stripping them from the Git history using git filter-branch commands would be helpful

Thanks for the tip. It was the documentation taking up all the space. I have cleared that history and now the repository is <800 KiB.

> The GitHub repository currently lacks the community guidelines, which is one of the JOSS review criteria. It could be a few sentences in the README.md welcoming contribution, and addition of CONTRIBUTING.md or issue templates would be a great plus, which can show up when someone opens up an issue.

As per your suggestion I have added community guidelines to the README. I have also added a
CONTRIBUTING.md file. And I have created issue templates.

Again, thank you @jongwook for the review.
