+++ CONTRIBUTING.md
@@ -7,14 +7,14 @@ Please feel free to suggest features, or additional stimuli you are interested i
 Or if you just want to show support, click the star button at the top of the page.
 
 This repository makes extensive use of GitHub actions.
-When you submit a pull request the continous integration will kick in to gear.
+When you submit a pull request the continuous integration will kick in to gear.
 The code in your pull request will then be automatically tested against several version of Julia,
 and on several operating systems to ensure that it doesn't break anything.
 
-Additionaly the documentation will be rebuilt based on your pull request.
+Additionally the documentation will be rebuilt based on your pull request.
 This will allow you to check that it renders correctly and your changes are realised.
 Further, a spell check will be performed on the pull request.
 
-Don't be disheartened if the continous integration results come back with a red cross intially.
+Don't be disheartened if the continuous integration results come back with a red cross initially.
 You can push further changes to fix any issues that arise.
 And if you get stuck just ping @rob-luke and we will solve the issues together.
