### 2021-08-19 #3613 @jongwook
@rob-luke Thanks for responding so fast, and I meant "40% of which is for tests" as a totally positive note as well!

I confirm that the git clone size has been reduced, and the community guidelines and the issue templates look great. Thanks!

@bmcfee All checkboxes marked off!
