### 2021-08-24 #3613 @rob-luke
> Hi @rob-luke. My first comment is that please add installation notes to the documentation.

As @jongwook notes "The documentation is very well-written as well, complete with installation instructions". Are the installation instructions are the first subtopic on the documentation page. See https://rob-luke.github.io/AuditoryStimuli.jl/dev/#Installation

This is common practice in Julia packages. What step is missing from these instructions that you would like me to add?
