### 2021-08-24 #3613 @vahid-sb
@rob-luke the dependencies for your package. I had to clone and install `MusicProcessing` from the disk. And it seems I have to do the same for `MP3` package too. Neither can be found in the registry, hence the example you have added cannot be run without extra steps. Also, the `MP3` package doc says it is not maintained anymore. Is there no other package in the registry that can load mp3 files?
