### 2021-08-24 #3613 @rob-luke
Hi @vahid-sb, where have you seen `MusicProcessing` in relation to this software package? It is unrelated to my software and review. I searched the repository for this term and it does not appear. Where did you get the impression it was required? Similarly the MP3 package is not utilised in my code or even discussed in the documentation.

The WAV package is utilised in the examples for this package but is not a core requirement. As such it is listed in the file `docs/Project.toml`, as per Julia standards. I am reasonably confident (but happy to be corrected) that all necessary packages are listed in the appropriate location as the continuous integration tests are passing and the documentation and the examples is also being built by the continuous integration each time code is modified.

I would be pleased to get to the cause of this confusion with you, how can I help? You can see the requirement including version restrictions at https://github.com/rob-luke/AuditoryStimuli.jl/blob/main/Project.toml and the extended requirements for the documentation (includes additional packages such as the documentation building software and useful tips for users) at https://github.com/rob-luke/AuditoryStimuli.jl/blob/main/docs/Project.toml
