### 2021-09-20 #3613 @vahid-sb
Hi @bmcfee
I haVe tested the code, run the example files and looked into the documentation. The code does what the documentation says it does and I find it comprehensive enough to be suitable for JOSS. As somebody who has worked on behavioural studies, I see there are some features that I might have needed and currently not available. But I do not find it reasonable to ask for all the features that various users may or may not need, at least in this stage of development. Hence, I'm ready to approve this software/paper in its current form to be published in JOSS.
