### 2021-09-20 #3613 @rob-luke
Wonderful!! Thanks @vahid-sb

Of course feel free to add any wish-list features as an issue on the repo. I plan to keep developing this software and adding more features as required.
