### 2021-09-21 #3613 @bmcfee
@rob-luke back to the bib formatting issues, can you try using `url={...}` instead of `howpublished={...}`?  It may not resolve the issue entirely, but it's the first thing to try.
