+++ paper.bib
@@ -30,7 +30,7 @@ title = {Pychoacoustics},
 year = {2012},
 publisher = {​GitHub},
 journal = {​GitHub repository},
-howpublished = {​https://github.com/sam81/pychoacoustics}
+url = {​https://github.com/sam81/pychoacoustics}
 }
 
 @article{psychopy2,
@@ -60,7 +60,7 @@ title = {Spatially Oriented Format for Acoustics (SOFA) API for Python},
 year = {2019},
 publisher = {​GitHub},
 journal = {​GitHub repository},
-howpublished = {​https://github.com/spatialaudio/python-sofa}
+url = {​https://github.com/spatialaudio/python-sofa}
 }
 
 @misc{unitful,
@@ -69,7 +69,7 @@ title = {Unitful.jl},
 year = {2016},
 publisher = {​GitHub},
 journal = {​GitHub repository},
-howpublished = {​https://github.com/PainterQubits/Unitful.jl}
+url = {​https://github.com/PainterQubits/Unitful.jl}
 }
 
 @misc{sampled-signals,
@@ -78,7 +78,7 @@ title = {SampledSignals.jl},
 year = {2016},
 publisher = {​GitHub},
 journal = {​GitHub repository},
-howpublished = {​https://github.com/JuliaAudio/SampledSignals.jl}
+url = {​https://github.com/JuliaAudio/SampledSignals.jl}
 }
 
 @article{gramfort2014mne,
