+++ paper.bib
@@ -58,27 +58,27 @@ doi = {10.25080/majora-7b98e3ed-003}
 author = {Jannika Lossner},
 title = {Spatially Oriented Format for Acoustics (SOFA) API for Python},
 year = {2019},
-publisher = {​GitHub},
-journal = {​GitHub repository},
-url = {​https://github.com/spatialaudio/python-sofa}
+publisher = {GitHub},
+journal = {GitHub repository},
+url = {https://github.com/spatialaudio/python-sofa}
 }
 
 @misc{unitful,
 author = {Andrew Keller},
 title = {Unitful.jl},
 year = {2016},
-publisher = {​GitHub},
-journal = {​GitHub repository},
-url = {​https://github.com/PainterQubits/Unitful.jl}
+publisher = {GitHub},
+journal = {GitHub repository},
+url = {https://github.com/PainterQubits/Unitful.jl}
 }
 
 @misc{sampled-signals,
 author = { Spencer Russell },
 title = {SampledSignals.jl},
 year = {2016},
-publisher = {​GitHub},
-journal = {​GitHub repository},
-url = {​https://github.com/JuliaAudio/SampledSignals.jl}
+publisher = {GitHub},
+journal = {GitHub repository},
+url = {https://github.com/JuliaAudio/SampledSignals.jl}
 }
 
 @article{gramfort2014mne,
