+++ paper.bib
@@ -28,9 +28,9 @@ journal = {Journal of Open Source Software}
 author = {Samuele Carcagno},
 title = {Pychoacoustics},
 year = {2012},
-publisher = {​GitHub},
-journal = {​GitHub repository},
-url = {​https://github.com/sam81/pychoacoustics}
+publisher = {GitHub},
+journal = {GitHub repository},
+url = {https://github.com/sam81/pychoacoustics}
 }
 
 @article{psychopy2,
