### 2021-09-22 #3613 @rob-luke
Actually I can see that the patent numbers aren't formatting correctly in the bibliography either. Please give me a few hours to try and fix this in the bib file.
