+++ paper.bib
@@ -132,14 +132,19 @@ publisher={Google Patents},
 note={US Patent 10,412,518}
 }
 
-@misc{sapozhnykov2020headset,
-title={Headset on ear state detection},
-author={Sapozhnykov, Vitaliy and Harvey, Thomas Ivan and Erfaniansaeedi, Nafiseh and Robert, LUKE},
-year={2020},
-month=oct # "~20",
-publisher={Google Patents},
-note={US Patent 10,812,889}
+
+@patent{sapozhnykov2020headset,
+ assignee = {Cirrus Logic},
+ author = {Sapozhnykov, Vitaliy and Harvey, Thomas Ivan and Erfaniansaeedi, Nafiseh and Luke, Robert},
+ month = {October},
+ monthfiled = {October},
+ nationality = {US},
+ note = {Patent},
+ number = {10,812,889},
+ title = {Headset on ear state detection},
+ year = {2021},
 }
+
 
 @article{luke2016kalman,
 title={Kalman filter based estimation of auditory steady state response parameters},
