+++ paper.bib
@@ -114,37 +114,41 @@ publisher={SIAM},
 doi = {10.1137/141000671}
 }
 
-@misc{sapozhnykov2019wind,
-title={Wind noise measurement},
-author={Sapozhnykov, Vitaliy and Harvey, Thomas Ivan and Robert, LUKE},
-year={2019},
-month=dec # "~10",
-publisher={Google Patents},
-note={US Patent 10,504,537}
+@patent{sapozhnykov2019wind,
+assignee = {Cirrus Logic},
+author = {Sapozhnykov, Vitaliy and Harvey, Thomas Ivan and Luke, Robert},
+month = {December},
+monthfiled = {December},
+nationality = {US},
+note = {Patent},
+number = {10,504,537},
+title = {Wind noise measurement},
+year = {2019},
 }
 
-@misc{robert2019blocked,
-title={Blocked microphone detection},
-author={Luke, Robert and Sapozhnykov, Vitaliy and Harvey, Thomas Ivan},
-year={2019},
-month=sep # "~10",
-publisher={Google Patents},
-note={US Patent 10,412,518}
+@patent{robert2019blocked,
+assignee = {Cirrus Logic},
+author = {Luke, Robert and Sapozhnykov, Vitaliy and Harvey, Thomas Ivan},
+month = {September},
+monthfiled = {September},
+nationality = {US},
+note = {Patent},
+number = {10,412,518},
+title = {Blocked microphone detection},
+year = {2020},
 }
-
 
 @patent{sapozhnykov2020headset,
- assignee = {Cirrus Logic},
- author = {Sapozhnykov, Vitaliy and Harvey, Thomas Ivan and Erfaniansaeedi, Nafiseh and Luke, Robert},
- month = {October},
- monthfiled = {October},
- nationality = {US},
- note = {Patent},
- number = {10,812,889},
- title = {Headset on ear state detection},
- year = {2021},
+assignee = {Cirrus Logic},
+author = {Sapozhnykov, Vitaliy and Harvey, Thomas Ivan and Erfaniansaeedi, Nafiseh and Luke, Robert},
+month = {October},
+monthfiled = {October},
+nationality = {US},
+note = {Patent},
+number = {10,812,889},
+title = {Headset on ear state detection},
+year = {2019},
 }
-
 
 @article{luke2016kalman,
 title={Kalman filter based estimation of auditory steady state response parameters},
