+++ paper.bib
@@ -124,6 +124,7 @@ note = {Patent},
 number = {10,504,537},
 title = {Wind noise measurement},
 year = {2019},
+url = {https://patents.google.com/patent/US10504537B2/en},
 }
 
 @patent{robert2019blocked,
@@ -136,6 +137,7 @@ note = {Patent},
 number = {10,412,518},
 title = {Blocked microphone detection},
 year = {2020},
+url = {https://patents.google.com/patent/US10412518B2/en},
 }
 
 @patent{sapozhnykov2020headset,
@@ -148,6 +150,7 @@ note = {Patent},
 number = {10,812,889},
 title = {Headset on ear state detection},
 year = {2019},
+url = {https://patents.google.com/patent/US10812889B2/en},
 }
 
 @article{luke2016kalman,
