### 2021-09-23 #3613 @rob-luke
Maybe I can help out our overworked friend Whedon.

There are no dois for the software references, so I added urls:
* https://github.com/sam81/pychoacoustics
* https://github.com/spatialaudio/python-sofa
* https://github.com/JuliaAudio/SampledSignals.jl

There are no dois for the patents, so I added URLs to the google patent pages:
* https://patents.google.com/patent/US10412518B2/en
* https://patents.google.com/patent/US10812889B2/en
* https://patents.google.com/patent/US10504537B2/en

I could not find a doi for:
* Fast music and audio processing using the Julia language by @jongwook
  https://www.aes.org/e-lib/browse.cfm?elib=18769
  There is a DOI for the whole proceedings (https://doi.org/10.17743/aesconf.2017.978-1-942220-15-2) should I use that?

I dont know much about the doi system, so please correct any misunderstanding I have.
