### 2021-09-23 #3613 @bmcfee
Thanks @rob-luke - I think that should be fine from the reference side.

You are right that there does not appear to be a DOI for that particular reference. I think it's okay to leave it as is; citing the entire proceedings would be confusing.

Some last minor comments on the article from my side:

- End of paragraph 1 is a sentence fragment
- Line 35: "users responses" => "users' responses" or "user's responses"?  Although, in context, "subject" might be more appropriate than "user" since the latter refers to the experimenter/researcher later in the article.  I'll defer to your judgment on that.
- Line 82: "researchs" => "researchers"

Otherwise, the last steps here are to push an up-to-date version of the software (in case anything has changed in response to the reviews); tag the github repo accordingly, and archive it to zenodo or figshare to get a DOI.  When you've done that, comment back with the software's DOI so we can link it to the paper.
