### 2021-09-23 149becf
Incorporate feedback from JOSS review
https://github.com/openjournals/joss-reviews/issues/3613#issuecomment-926027674
