+++ README.md
@@ -2,6 +2,7 @@
 
 ![Tests](https://github.com/rob-luke/AuditoryStimuli.jl/workflows/Tests/badge.svg)
 [![codecov.io](http://codecov.io/github/rob-luke/AuditoryStimuli.jl/coverage.svg?branch=master)](http://codecov.io/github/rob-luke/AuditoryStimuli.jl?branch=master)
+[![DOI](https://zenodo.org/badge/141233777.svg)](https://zenodo.org/badge/latestdoi/141233777)
 
 Generate auditory stimuli for real-time applications.  Specifically, stimuli that are used in auditory research.
 
