### 2021-09-23 #3613 @rob-luke
Thanks for the review @bmcfee. I removed the entire sentence fragment as the same content was addressed in the final paragraph. I changed `users responses` to `participant's responses`, and fixed the typo you found. Thanks for the extra set of eyes!

I just released a new version of the software. And also created a Zenodo doi `10.5281/zenodo.5525443`.
