### 2021-09-24 #3613 @kthyng
Hi @rob-luke! Final steps to wrap up here. Can you update the Zenodo archive metadata so the title and author list exactly match the JOSS paper?
