### 2021-09-24 #3613 @kthyng
Paper comments:
* first page, "existing tools such as (Carcagno 2012)" — this reference should be inline, not parenthetical. You can make this change by removing the [] around the reference.
* final sentence in that paragraph — don't start a sentence with "i.e", at least write out as 'That is' or change wording.
* Same parenthetical problem with a reference in the "Notes" section.
* fix capitalization in your references, for example "python" and "julia" should be capitalized. You can preserve capitalization in your .bib file by putting {} around strings.
