+++ paper.bib
@@ -1,5 +1,5 @@
 @article{luke2021analysis,
-title={Analysis methods for measuring passive auditory fNIRS responses generated by a block-design paradigm},
+title={Analysis methods for measuring passive auditory f{NIRS} responses generated by a block-design paradigm},
 author={Luke, Robert and Larson, Eric and Shader, Maureen J and Innes-Brown, Hamish and Van Yper, Lindsey and Lee, Adrian KC and Sowman, Paul F and McAlpine, David},
 journal={Neurophotonics},
 volume={8},
@@ -19,7 +19,7 @@ volume = {6},
 number = {62},
 pages = {3284},
 author = {Marc Schönwiesner and Ole Bialas},
-title = {s(ound)lab: An easy to learn Python package for designing and running psychoacoustic experiments.},
+title = {s(ound)lab: An easy to learn {P}ython package for designing and running psychoacoustic experiments.},
 journal = {Journal of Open Source Software}
 }
 
@@ -34,7 +34,7 @@ url = {https://github.com/sam81/pychoacoustics}
 }
 
 @article{psychopy2,
-title = {PsychoPy2: Experiments in behavior made easy},
+title = {Psycho{P}y2: Experiments in behavior made easy},
 volume = {51},
 issn = {1554-3528},
 doi = {10.3758/s13428-018-01193-y},
@@ -46,7 +46,7 @@ pages = {195--203}
 }
 
 @inproceedings{mcfee2015librosa,
-title={librosa: Audio and music signal analysis in python},
+title={librosa: {A}udio and music signal analysis in {P}ython},
 author={McFee, Brian and Raffel, Colin and Liang, Dawen and Ellis, Daniel PW and McVicar, Matt and Battenberg, Eric and Nieto, Oriol},
 booktitle={Proceedings of the 14th python in science conference},
 volume={8},
@@ -56,7 +56,7 @@ doi = {10.25080/majora-7b98e3ed-003}
 
 @misc{python-sofa,
 author = {Jannika Lossner},
-title = {Spatially Oriented Format for Acoustics (SOFA) API for Python},
+title = {Spatially {O}riented {F}ormat for {A}coustics {(SOFA)} {API} for {P}ython},
 year = {2019},
 publisher = {GitHub},
 journal = {GitHub repository},
@@ -82,7 +82,7 @@ url = {https://github.com/JuliaAudio/SampledSignals.jl}
 }
 
 @article{gramfort2014mne,
-title={MNE software for processing MEG and EEG data},
+title={{MNE} software for processing {MEG} and {EEG} data},
 author={Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A and Strohmeier, Daniel and Brodbeck, Christian and Parkkonen, Lauri and H{\"a}m{\"a}l{\"a}inen, Matti S},
 journal={Neuroimage},
 volume={86},
@@ -93,7 +93,7 @@ doi = {10.1016/j.neuroimage.2013.10.027}
 }
 
 @article{oostenveld2011fieldtrip,
-title={FieldTrip: open source software for advanced analysis of MEG, EEG, and invasive electrophysiological data},
+title={Field{T}rip: open source software for advanced analysis of {MEG}, {EEG}, and invasive electrophysiological data},
 author={Oostenveld, Robert and Fries, Pascal and Maris, Eric and Schoffelen, Jan-Mathijs},
 journal={Computational intelligence and neuroscience},
 volume={2011},
@@ -180,7 +180,7 @@ Volume = {172},
 Year = {2018}}
 
 @inproceedings{kim2017fast,
-title={Fast music and audio processing using the Julia language},
+title={Fast music and audio processing using the {J}ulia language},
[…]
