### 2021-09-24 #3613 @rob-luke
Thanks for carefully checking the details @kthyng. I have updated the zenodo title and made the changes to the manuscript and bib that you suggested. Please let me know if there is anything else I should do. Cheers
